+++ celescope/flv_CR/assemble.py
@@ -1,5 +1,5 @@
 import os
-from celescope.fl_vdj_CR.VDJ_Mixin import VDJ_Mixin, get_opts_VDJ_Mixin
+from celescope.flv_CR.VDJ_Mixin import VDJ_Mixin, get_opts_VDJ_Mixin
 from celescope.tools.step import s_common