+++ celescope/tools/__init__.py
@@ -1,17 +1,17 @@
 # barcode
 __PATTERN_DICT__ = {
+    'auto': None,
+    'scopeV1': 'C12U8T18',
     'scopeV2.0.0': 'C8L16C8L16C8U8T18',
     'scopeV2.0.1': 'C8L16C8L16C8L1U8T18',
     'scopeV2.1.0': 'C8L16C8L16C8U12T18',
     'scopeV2.1.1': 'C8L16C8L16C8L1U12T18',
     'scopeV2.2.1': 'C8L16C8L16C8L1U12T18',
     'scopeV3.0.1': 'C9L16C9L16C9L1U12T18',
-    'customized': None,
-    'auto': None,
-    'scopeV1': 'C12U8T18',
     # flv_rna is actually U9L16, but the last 
     'flv_rna': 'C8L16C8L16C8U9L6',
     'flv' : 'U9C8L16C8L16C8',
+    'customized': None,
 }